+++ HISTORY.md
@@ -60,3 +60,11 @@
  - BREAKING CHANGE: change commands to `extensionRelease`, `npmRelease` and `repoRelease` and remove the commands `oldWayRelease`, `createRelease` and `prepareRelease`
  - BREAKING CHANGE: next version is computed from conventional commits 
  - BREAKING CHANGE: `extensionRelease` doesn't update version in the manifest anymore
+
+
+## [Version 2.1.0] 
+
+ - Support non interactive mode. (see `--no-interactive` option).
+ - Writing to the config is configurable. (see `--no-write` option). 
+ - can read the `GITHUB_TOKEN` from env.
+